fix minutes return

+++ app/combine/create_table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS combines(
   nb_match_win float,
   h2h float,
-  minutes float,
+  minutes integer,
   id serial primary key not null
 );
